+++ lib/z_drag_and_drop/drag_and_drop_item_target.dart
@@ -9,13 +9,12 @@ class DragAndDropItemTarget extends StatefulWidget {
   final DragAndDropBuilderParameters parameters;
   final OnItemDropOnLastTarget onReorderOrAdd;
 
-  const DragAndDropItemTarget({
-    required this.child,
-    required this.onReorderOrAdd,
-    required this.parameters,
-    this.parent,
-    super.key
-  });
+  const DragAndDropItemTarget(
+      {required this.child,
+      required this.onReorderOrAdd,
+      required this.parameters,
+      this.parent,
+      super.key});
 
   @override
   State<StatefulWidget> createState() => _DragAndDropItemTarget();
@@ -28,24 +27,25 @@ class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
   @override
   Widget build(BuildContext context) {
     return Stack(
-      children: [
+      children: <Widget>[
         Column(
           crossAxisAlignment: widget.parameters.verticalAlignment,
-          children: [
+          children: <Widget>[
             AnimatedSize(
+              duration: Duration(
+                  milliseconds: widget.parameters.itemSizeAnimationDuration),
               alignment: Alignment.bottomCenter,
-              duration: Duration(milliseconds: widget.parameters.itemSizeAnimationDuration),
-
-              child: _hoveredDraggable == null ? Container() : Opacity(
-                opacity: widget.parameters.itemGhostOpacity,
-                child: widget.parameters.itemGhost ?? _hoveredDraggable!.child,
-              ),
+              child: _hoveredDraggable != null
+                  ? Opacity(
+                      opacity: widget.parameters.itemGhostOpacity,
+                      child: widget.parameters.itemGhost ??
+                          _hoveredDraggable!.child,
+                    )
+                  : Container(),
             ),
-            
             widget.child,
           ],
         ),
-
         Positioned.fill(
           child: DragTarget<DragAndDropItem>(
             builder: (context, candidateData, rejectedData) {
@@ -55,16 +55,21 @@ class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
             onWillAcceptWithDetails: (details) {
               bool accept = true;
               if (widget.parameters.itemTargetOnWillAccept != null) {
-                accept = widget.parameters.itemTargetOnWillAccept!(details.data, widget);
+                accept =
+                    widget.parameters.itemTargetOnWillAccept!(details.data, widget);
               }
               if (accept && mounted) {
-                setState(() => _hoveredDraggable = details.data);
+                setState(() {
+                  _hoveredDraggable = details.data;
+                });
               }
               return accept;
             },
             onLeave: (data) {
               if (mounted) {
-                setState(() => _hoveredDraggable = null);
+                setState(() {
+                  _hoveredDraggable = null;
+                });
               }
             },
             onAcceptWithDetails: (details) {