+++ lib/z_drag_and_drop/drag_and_drop_list.dart
@@ -7,28 +7,53 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 class DragAndDropList implements DragAndDropListInterface {
-  final double? height;
+  /// The widget that is displayed at the top of the list.
   final Widget? header;
+
+  /// The widget that is displayed at the bottom of the list.
   final Widget? footer;
+
+  /// The widget that is displayed to the left of the list.
   final Widget? leftSide;
+
+  /// The widget that is displayed to the right of the list.
   final Widget? rightSide;
+
+  /// The widget to be displayed when a list is empty.
+  /// If this is not null, it will override that set in [DragAndDropLists.contentsWhenEmpty].
   final Widget? contentsWhenEmpty;
+
+  /// The widget to be displayed as the last element in the list that will accept
+  /// a dragged item.
   final Widget? lastTarget;
+
+  /// The decoration displayed around a list.
+  /// If this is not null, it will override that set in [DragAndDropLists.listDecoration].
   final Decoration? decoration;
+
+  /// The vertical alignment of the contents in this list.
+  /// If this is not null, it will override that set in [DragAndDropLists.verticalAlignment].
   final CrossAxisAlignment verticalAlignment;
+
+  /// The horizontal alignment of the contents in this list.
+  /// If this is not null, it will override that set in [DragAndDropLists.horizontalAlignment].
   final MainAxisAlignment horizontalAlignment;
 
+  /// The child elements that will be contained in this list.
+  /// It is possible to not provide any children when an empty list is desired.
   @override
   final List<DragAndDropItem> children;
 
+  /// Whether or not this item can be dragged.
+  /// Set to true if it can be reordered.
+  /// Set to false if it must remain fixed.
   @override
   final bool canDrag;
+  @override
   final Key? key;
-
   DragAndDropList({
     required this.children,
     this.key,
-    this.height,
     this.header,
     this.footer,
     this.leftSide,
@@ -44,11 +69,9 @@ class DragAndDropList implements DragAndDropListInterface {
   @override
   Widget generateWidget(DragAndDropBuilderParameters params) {
     var contents = <Widget>[];
-
     if (header != null) {
       contents.add(Flexible(child: header!));
     }
-
     Widget intrinsicHeight = IntrinsicHeight(
       child: Row(
         mainAxisAlignment: horizontalAlignment,
@@ -57,21 +80,18 @@ class DragAndDropList implements DragAndDropListInterface {
         children: _generateDragAndDropListInnerContents(params),
       ),
     );
-
     if (params.axis == Axis.horizontal) {
       intrinsicHeight = SizedBox(
         width: params.listWidth,
         child: intrinsicHeight,
       );
     }
-
     if (params.listInnerDecoration != null) {
       intrinsicHeight = Container(
         decoration: params.listInnerDecoration,
         child: intrinsicHeight,
       );
     }
-    
     contents.add(intrinsicHeight);
 
     if (footer != null) {
@@ -80,10 +100,10 @@ class DragAndDropList implements DragAndDropListInterface {
 
     return Container(
       key: key,
+      width: params.axis == Axis.vertical
+          ? double.infinity
+          : params.listWidth - params.listPadding!.horizontal,
       decoration: decoration ?? params.listDecoration,
-      width: params.axis == Axis.vertical ? double.infinity : params.listWidth - params.listPadding!.horizontal,
-      height: params.axis == Axis.vertical ? null : 500,
-
       child: Column(
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: verticalAlignment,
@@ -92,40 +112,38 @@ class DragAndDropList implements DragAndDropListInterface {
     );
   }
 
-  List<Widget> _generateDragAndDropListInnerContents(DragAndDropBuilderParameters parameters) {
+  List<Widget> _generateDragAndDropListInnerContents(
+      DragAndDropBuilderParameters parameters) {
     var contents = <Widget>[];
-
     if (leftSide != null) {
       contents.add(leftSide!);
     }
-    
     if (children.isNotEmpty) {
-      List<Widget> allChildren = [];
+      List<Widget> allChildren = <Widget>[];
       if (parameters.addLastItemTargetHeightToTop) {
         allChildren.add(Padding(
           padding: EdgeInsets.only(top: parameters.lastItemTargetHeight),
         ));
       }
-
       for (int i = 0; i < children.length; i++) {
         allChildren.add(DragAndDropItemWrapper(
           key: children[i].key,
           child: children[i],
           parameters: parameters,
         ));
-
         if (parameters.itemDivider != null && i < children.length - 1) {
           allChildren.add(parameters.itemDivider!);
         }
       }
-
       allChildren.add(DragAndDropItemTarget(
         parent: this,
         parameters: parameters,
         onReorderOrAdd: parameters.onItemDropOnLastTarget!,
-        child: lastTarget ?? Container(height: parameters.lastItemTargetHeight),
+        child: lastTarget ??
+            Container(
+              height: parameters.lastItemTargetHeight,
+            ),
       ));
-
       contents.add(
         Expanded(
           child: SingleChildScrollView(
@@ -145,19 +163,22 @@ class DragAndDropList implements DragAndDropListInterface {
             physics: const NeverScrollableScrollPhysics(),
             child: Column(
               mainAxisSize: MainAxisSize.max,
-              children: [
-                contentsWhenEmpty ?? const Text(
-                  'Empty list',
-                  style: TextStyle(
-                    fontStyle: FontStyle.italic,
-                  ),
-                ),
-
+              children: <Widget>[
+                contentsWhenEmpty ??
+                    const Text(
+                      'Empty list',
+                      style: TextStyle(
+                        fontStyle: FontStyle.italic,
+                      ),
+                    ),
                 DragAndDropItemTarget(
                   parent: this,
                   parameters: parameters,
                   onReorderOrAdd: parameters.onItemDropOnLastTarget!,
-                  child: lastTarget ?? Container(height: parameters.lastItemTargetHeight),
+                  child: lastTarget ??
+                      Container(
+                        height: parameters.lastItemTargetHeight,
+                      ),
                 ),
               ],
             ),
@@ -165,11 +186,9 @@ class DragAndDropList implements DragAndDropListInterface {
         ),
       );
     }
-    
     if (rightSide != null) {
       contents.add(rightSide!);
     }
-
     return contents;
   }
 }