+++ lib/z_drag_and_drop/drag_and_drop_list_wrapper.dart
@@ -117,7 +117,7 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
       draggable = dragAndDropListContents;
     }
 
-    var rowOrColumnChildren = [
+    var rowOrColumnChildren = <Widget>[
       AnimatedSize(
         duration:
             Duration(milliseconds: widget.parameters.listSizeAnimationDuration),
@@ -149,16 +149,15 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
     ];
 
     var stack = Stack(
-      children: [
+      children: <Widget>[
         widget.parameters.axis == Axis.vertical
-          ? Column(
-              children: rowOrColumnChildren,
-            )
-          : Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: rowOrColumnChildren,
-            ),
-            
+            ? Column(
+                children: rowOrColumnChildren,
+              )
+            : Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: rowOrColumnChildren,
+              ),
         Positioned.fill(
           child: DragTarget<DragAndDropListInterface>(
             builder: (context, candidateData, rejectedData) {
@@ -168,25 +167,30 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
             onWillAcceptWithDetails: (details) {
               bool accept = true;
               if (widget.parameters.listOnWillAccept != null) {
-                accept = widget.parameters.listOnWillAccept!(details.data, widget.dragAndDropList);
+                accept = widget.parameters.listOnWillAccept!(
+                    details.data, widget.dragAndDropList);
               }
-              
               if (accept && mounted) {
-                setState(() => _hoveredDraggable = details.data);
+                setState(() {
+                  _hoveredDraggable = details.data;
+                });
               }
               return accept;
             },
             onLeave: (data) {
               if (_hoveredDraggable != null) {
                 if (mounted) {
-                  setState(() => _hoveredDraggable = null);
+                  setState(() {
+                    _hoveredDraggable = null;
+                  });
                 }
               }
             },
             onAcceptWithDetails: (details) {
               if (mounted) {
                 setState(() {
-                  widget.parameters.onListReordered!(details.data, widget.dragAndDropList);
+                  widget.parameters.onListReordered!(
+                      details.data, widget.dragAndDropList);
                   _hoveredDraggable = null;
                 });
               }
@@ -197,25 +201,26 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
     );
 
     Widget toReturn = stack;
-
     if (widget.parameters.listPadding != null) {
       toReturn = Padding(
         padding: widget.parameters.listPadding!,
         child: stack,
       );
     }
-
     if (widget.parameters.axis == Axis.horizontal &&
         !widget.parameters.disableScrolling) {
       toReturn = SingleChildScrollView(
-        child: Container(child: toReturn),
+        child: Container(
+          child: toReturn,
+        ),
       );
     }
 
     return toReturn;
   }
 
-  Material buildFeedbackWithHandle(Widget dragAndDropListContents, Widget dragHandle) {
+  Material buildFeedbackWithHandle(
+      Widget dragAndDropListContents, Widget dragHandle) {
     return Material(
       color: Colors.transparent,
       child: Container(
@@ -228,17 +233,17 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
                 textDirection: Directionality.of(context),
                 child: dragAndDropListContents,
               ),
-
               Positioned(
                 right: widget.parameters.listDragHandle!.onLeft ? null : 0,
                 left: widget.parameters.listDragHandle!.onLeft ? 0 : null,
-                top: widget.parameters.listDragHandle!.verticalAlignment == DragHandleVerticalAlignment.bottom
+                top: widget.parameters.listDragHandle!.verticalAlignment ==
+                        DragHandleVerticalAlignment.bottom
                     ? null
                     : 0,
-                bottom: widget.parameters.listDragHandle!.verticalAlignment == DragHandleVerticalAlignment.top
+                bottom: widget.parameters.listDragHandle!.verticalAlignment ==
+                        DragHandleVerticalAlignment.top
                     ? null
                     : 0,
-
                 child: dragHandle,
               ),
             ],
@@ -248,12 +253,14 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
     );
   }
 
-  SizedBox buildFeedbackWithoutHandle(BuildContext context, Widget dragAndDropListContents) {
+  SizedBox buildFeedbackWithoutHandle(
+      BuildContext context, Widget dragAndDropListContents) {
     return SizedBox(
       width: widget.parameters.axis == Axis.vertical
-        ? (widget.parameters.listDraggingWidth ?? MediaQuery.of(context).size.width)
-        : (widget.parameters.listDraggingWidth ?? widget.parameters.listWidth),
-
+          ? (widget.parameters.listDraggingWidth ??
+              MediaQuery.of(context).size.width)
+          : (widget.parameters.listDraggingWidth ??
+              widget.parameters.listWidth),
       child: Material(
         color: Colors.transparent,
         child: Container(
@@ -268,9 +275,10 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
   }
 
   Axis? draggableAxis() {
-    return widget.parameters.axis == Axis.vertical && widget.parameters.constrainDraggingAxis
-      ? Axis.vertical
-      : null;
+    return widget.parameters.axis == Axis.vertical &&
+            widget.parameters.constrainDraggingAxis
+        ? Axis.vertical
+        : null;
   }
 
   double _dragHandleDistanceFromTop() {
@@ -289,13 +297,13 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
   Offset _feedbackContainerOffset() {
     double xOffset;
     double yOffset;
-
     if (widget.parameters.listDragHandle!.onLeft) {
       xOffset = 0;
     } else {
       xOffset = -_containerSize.width + _dragHandleSize.width;
     }
-    if (widget.parameters.listDragHandle!.verticalAlignment == DragHandleVerticalAlignment.bottom) {
+    if (widget.parameters.listDragHandle!.verticalAlignment ==
+        DragHandleVerticalAlignment.bottom) {
       yOffset = -_containerSize.height + _dragHandleSize.width;
     } else {
       yOffset = 0;
@@ -306,10 +314,12 @@ class _DragAndDropListWrapper extends State<DragAndDropListWrapper>
 
   void _setDragging(bool dragging) {
     if (_dragging != dragging && mounted) {
-      setState(() => _dragging = dragging);
-
+      setState(() {
+        _dragging = dragging;
+      });
       if (widget.parameters.onListDraggingChanged != null) {
-        widget.parameters.onListDraggingChanged!(widget.dragAndDropList, dragging);
+        widget.parameters.onListDraggingChanged!(
+            widget.dragAndDropList, dragging);
       }
     }
   }