+++ lib/z_drag_and_drop/drag_handle.dart
@@ -7,9 +7,14 @@ enum DragHandleVerticalAlignment {
 }
 
 class DragHandle extends StatelessWidget {
+  /// Set the drag handle to be on the left side instead of the default right side
   final bool onLeft;
-  final Widget child;
+
+  /// Align the list drag handle to the top, center, or bottom
   final DragHandleVerticalAlignment verticalAlignment;
+
+  /// Child widget to displaying the drag handle
+  final Widget child;
 
   const DragHandle({
     super.key,