+++ lib/z_drag_and_drop/programmatic_expansion_tile.dart
@@ -7,7 +7,25 @@ import 'package:flutter/scheduler.dart';
 
 const Duration _kExpand = Duration(milliseconds: 200);
 
+/// A single-line [ListTile] with a trailing button that expands or collapses
+/// the tile to reveal or hide the [children].
+///
+/// This widget is typically used with [ListView] to create an
+/// "expand / collapse" list entry. When used with scrolling widgets like
+/// [ListView], a unique [PageStorageKey] must be specified to enable the
+/// [ProgrammaticExpansionTile] to save and restore its expanded state when it is scrolled
+/// in and out of view.
+///
+/// See also:
+///
+///  * [ListTile], useful for creating expansion tile [children] when the
+///    expansion tile represents a sublist.
+///  * The "Expand/collapse" section of
+///    <https://material.io/guidelines/components/lists-controls.html>.
 class ProgrammaticExpansionTile extends StatefulWidget {
+  /// Creates a single-line [ListTile] with a trailing button that expands or collapses
+  /// the tile to reveal or hide the [children]. The [initiallyExpanded] property must
+  /// be non-null.
   const ProgrammaticExpansionTile({
     required Key key,
     required this.listKey,
@@ -24,25 +42,64 @@ class ProgrammaticExpansionTile extends StatefulWidget {
   }) : super(key: key);
 
   final Key listKey;
+
+  /// A widget to display before the title.
+  ///
+  /// Typically a [CircleAvatar] widget.
   final Widget? leading;
+
+  /// The primary content of the list item.
+  ///
+  /// Typically a [Text] widget.
   final Widget? title;
+
+  /// Additional content displayed below the title.
+  ///
+  /// Typically a [Text] widget.
   final Widget? subtitle;
+
+  /// Additional content displayed below the title.
+  ///
+  /// Typically a [Text] widget.
   final bool isThreeLine;
+
+  /// Called when the tile expands or collapses.
+  ///
+  /// When the tile starts expanding, this function is called with the value
+  /// true. When the tile starts collapsing, this function is called with
+  /// the value false.
   final ValueChanged<bool>? onExpansionChanged;
+
+  /// The widgets that are displayed when the tile expands.
+  ///
+  /// Typically [ListTile] widgets.
   final List<Widget?> children;
+
+  /// The color to display behind the sublist when expanded.
   final Color? backgroundColor;
+
+  /// A widget to display instead of a rotating arrow icon.
   final Widget? trailing;
+
+  /// Specifies if the list tile is initially expanded (true) or collapsed (false, the default).
   final bool initiallyExpanded;
+
+  /// Disable to borders displayed at the top and bottom when expanded
   final bool disableTopAndBottomBorders;
 
   @override
-  ProgrammaticExpansionTileState createState() => ProgrammaticExpansionTileState();
+  ProgrammaticExpansionTileState createState() =>
+      ProgrammaticExpansionTileState();
 }
 
-class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> with SingleTickerProviderStateMixin {
-  static final Animatable<double> _easeOutTween = CurveTween(curve: Curves.easeOut);
-  static final Animatable<double> _easeInTween = CurveTween(curve: Curves.easeIn);
-  static final Animatable<double> _halfTween = Tween<double>(begin: 0.0, end: 0.5);
+class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile>
+    with SingleTickerProviderStateMixin {
+  static final Animatable<double> _easeOutTween =
+      CurveTween(curve: Curves.easeOut);
+  static final Animatable<double> _easeInTween =
+      CurveTween(curve: Curves.easeIn);
+  static final Animatable<double> _halfTween =
+      Tween<double>(begin: 0.0, end: 0.5);
 
   final ColorTween _borderColorTween = ColorTween();
   final ColorTween _headerColorTween = ColorTween();
@@ -68,15 +125,19 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
     _borderColor = _controller.drive(_borderColorTween.chain(_easeOutTween));
     _headerColor = _controller.drive(_headerColorTween.chain(_easeInTween));
     _iconColor = _controller.drive(_iconColorTween.chain(_easeInTween));
-    _backgroundColor = _controller.drive(_backgroundColorTween.chain(_easeOutTween));
+    _backgroundColor =
+        _controller.drive(_backgroundColorTween.chain(_easeOutTween));
 
     _isExpanded = PageStorage.of(context)
-                             .readState(context, identifier: widget.listKey) as bool? ?? widget.initiallyExpanded;
-
+            .readState(context, identifier: widget.listKey) as bool? ??
+        widget.initiallyExpanded;
     if (_isExpanded) _controller.value = 1.0;
 
+    // Schedule the notification that widget has changed for after init
+    // to ensure that the parent widget maintains the correct state
     SchedulerBinding.instance.addPostFrameCallback((Duration duration) {
-      if (widget.onExpansionChanged != null && _isExpanded != widget.initiallyExpanded) {
+      if (widget.onExpansionChanged != null &&
+          _isExpanded != widget.initiallyExpanded) {
         widget.onExpansionChanged!(_isExpanded);
       }
     });
@@ -88,27 +149,34 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
     super.dispose();
   }
 
-  void expand() => _setExpanded(true);
+  void expand() {
+    _setExpanded(true);
+  }
 
-  void collapse() => _setExpanded(false);
+  void collapse() {
+    _setExpanded(false);
+  }
 
-  void toggle() =>  _setExpanded(!_isExpanded);
+  void toggle() {
+    _setExpanded(!_isExpanded);
+  }
 
   void _setExpanded(bool expanded) {
     if (_isExpanded != expanded) {
       setState(() {
         _isExpanded = expanded;
-
-        if (_isExpanded) { _controller.forward(); } 
-        else {
+        if (_isExpanded) {
+          _controller.forward();
+        } else {
           _controller.reverse().then<void>((void value) {
             if (!mounted) return;
-            setState(() {});
+            setState(() {
+              // Rebuild without widget.children.
+            });
           });
         }
-
         PageStorage.of(context)
-                   .writeState(context, _isExpanded, identifier: widget.listKey);
+            .writeState(context, _isExpanded, identifier: widget.listKey);
       });
       if (widget.onExpansionChanged != null) {
         widget.onExpansionChanged!(_isExpanded);
@@ -123,14 +191,16 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
     return Container(
       decoration: BoxDecoration(
         color: _backgroundColor.value ?? Colors.transparent,
-        border: !setBorder ? null : Border(
-          top: BorderSide(color: borderSideColor),
-          bottom: BorderSide(color: borderSideColor),
-        ),
+        border: setBorder
+            ? Border(
+                top: BorderSide(color: borderSideColor),
+                bottom: BorderSide(color: borderSideColor),
+              )
+            : null,
       ),
       child: Column(
         mainAxisSize: MainAxisSize.min,
-        children: [
+        children: <Widget>[
           ListTileTheme.merge(
             iconColor: _iconColor.value,
             textColor: _headerColor.value,
@@ -140,13 +210,13 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
               title: widget.title,
               subtitle: widget.subtitle,
               isThreeLine: widget.isThreeLine,
-              trailing: widget.trailing ?? RotationTransition(
-                turns: _iconTurns,
-                child: const Icon(Icons.expand_more),
-              ),
+              trailing: widget.trailing ??
+                  RotationTransition(
+                    turns: _iconTurns,
+                    child: const Icon(Icons.expand_more),
+                  ),
             ),
           ),
-
           ClipRect(
             child: Align(
               heightFactor: _heightFactor.value,
@@ -161,7 +231,6 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
   @override
   void didChangeDependencies() {
     final ThemeData theme = Theme.of(context);
-
     _borderColorTween.end = theme.dividerColor;
     _headerColorTween
       ..begin = theme.textTheme.titleMedium!.color
@@ -170,7 +239,6 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile> wi
       ..begin = theme.unselectedWidgetColor
       ..end = theme.colorScheme.secondary;
     _backgroundColorTween.end = widget.backgroundColor;
-    
     super.didChangeDependencies();
   }